BUG: Create script to verify email addresses that should be verified.
